NO-ISSUE: re-organize agent tests
